ltx: split one-stage pipeline; dedup test helpers into utils/ltx
Test names now follow the Lightricks stage vocabulary (one_stage/two_stages/
distilled) so future pipeline ports drop in cleanly. create_pipeline is a
classmethod resolving the concrete subclass via cls, dropping the lazy import
and per-subclass overrides.
